+++ Acquirer/index.ts
@@ -1,12 +1,11 @@
 import { Bin as AcquirerBin } from "./Bin"
 import { Protocol as AcquirerProtocol } from "./Protocol"
-import { Creatable as CreatableAcquirer } from "./Creatable"
 
 export interface Acquirer {
 	id?: string
-	protocol: AcquirerProtocol
-	url: string
-	key: string
+	protocol?: AcquirerProtocol
+	url?: string
+	key?: string
 	bin?: AcquirerBin
 }
 
@@ -15,9 +14,9 @@ export namespace Acquirer {
 		return (
 			typeof value == "object" &&
 			(value.id == undefined || typeof value.id == "string") &&
-			AcquirerProtocol.is(value.protocol) &&
-			typeof value.url == "string" &&
-			typeof value.key == "string" &&
+			(value.protocol == undefined || AcquirerProtocol.is(value.protocol)) &&
+			(value.url == undefined || typeof value.url == "string") &&
+			(value.key == undefined || typeof value.key == "string") &&
 			(value.bin == undefined || AcquirerBin.is(value.bin))
 		)
 	}
@@ -25,9 +24,5 @@ export namespace Acquirer {
 	export namespace Bin {
 		export const is = AcquirerBin.is
 		export const flaw = AcquirerBin.flaw
-	}
-	export type Acquirer = CreatableAcquirer
-	export namespace Acquirer {
-		export const is = CreatableAcquirer.is
 	}
 }
