+++ Merchant/index.ts
@@ -2,7 +2,7 @@ import * as gracely from "gracely"
 import * as authly from "authly"
 import { Key as MerchantKey } from "./Key"
 import * as MerchantV1 from "./V1"
-import { Configuration as MerchantConfiguration } from "./Configuration"
+import { Card as MerchantConfiguration } from "./Card"
 import { Emv3d as MerchantEmv3d } from "./Emv3d"
 
 export interface Merchant {
@@ -40,34 +40,10 @@ export namespace Merchant {
 					  ].filter(gracely.Flaw.is) as gracely.Flaw[]),
 		}
 	}
-	export type Configuration = MerchantConfiguration
-	export namespace Configuration {
+	export type Card = MerchantConfiguration
+	export namespace Card {
 		export const is = MerchantConfiguration.is
 		export const flaw = MerchantConfiguration.flaw
-		export type Card = MerchantConfiguration.Card
-		export namespace Card {
-			export const is = MerchantConfiguration.Card.is
-			export const flaw = MerchantConfiguration.Card.flaw
-			export type Creatable = MerchantConfiguration.Card.Creatable
-			export namespace Creatable {
-				export const is = MerchantConfiguration.Card.Creatable.is
-			}
-			export type KeyInfo = MerchantConfiguration.Card.KeyInfo
-			export namespace KeyInfo {
-				export const is = MerchantConfiguration.Card.KeyInfo.is
-				export const flaw = MerchantConfiguration.Card.KeyInfo.flaw
-			}
-		}
-		export type Override = MerchantConfiguration.Override
-		export namespace Override {
-			export const is = MerchantConfiguration.Override.is
-			export const flaw = MerchantConfiguration.Override.flaw
-		}
-		export type KeyInfo = MerchantConfiguration.KeyInfo
-		export namespace KeyInfo {
-			export const is = MerchantConfiguration.KeyInfo.is
-			export const flaw = MerchantConfiguration.KeyInfo.flaw
-		}
 	}
 	export type Emv3d = MerchantEmv3d
 	export namespace Emv3d {
@@ -89,16 +65,10 @@ export namespace Merchant {
 		export const is = MerchantKey.is
 		export const flaw = MerchantKey.flaw
 		export const upgrade = MerchantKey.upgrade
+		export const unpack = MerchantKey.unpack
 		export type Audience = MerchantKey.Audience
 		export namespace Audience {
 			export const is = MerchantKey.Audience.is
-		}
-		export type KeyInfo = MerchantKey.KeyInfo
-		export namespace KeyInfo {
-			export const is = MerchantKey.KeyInfo.is
-			export const flaw = MerchantKey.KeyInfo.flaw
-			export const unpack = MerchantKey.KeyInfo.unpack
-			export const upgrade = MerchantKey.KeyInfo.upgrade
 		}
 	}
 	export namespace V1 {
@@ -106,11 +76,6 @@ export namespace Merchant {
 		export namespace Key {
 			export const is = MerchantV1.Key.is
 			export const flaw = MerchantV1.Key.flaw
-			export type KeyInfo = MerchantV1.Key.KeyInfo
-			export namespace KeyInfo {
-				export const is = MerchantV1.Key.KeyInfo.is
-				export const flaw = MerchantV1.Key.KeyInfo.flaw
-			}
 		}
 	}
 }
